criu: RFC: added support for remote lazy restore
This is the second attempt to implement lazy restore between two hosts.
The memory pages which should be restored lazily (on demand) using
userfaultfd are not on the system were the process is restored and
therefore the required pages need to be transferred over the network.

In the first attempt the logic (the brain) of the restore process was
running on the source system and the uffd daemon on the destination
system was just forwarding the pages and page request to/from the
userfaultfd.

This new implementation is now based on the existing page-server and its
protocols (plus additions (see previous patch)). The logic which pages
are transferred is now in the uffd daemon on the restore side.

The current implementation is only the first step and still requires
additional work, but it offers a working remote lazy-restore
implementation.

Following commands are necessary to use this implementation:

Source system:

 criu page-server -D /tmp/4/ --port 27

Destination system:

 criu lazy-pages -D /tmp/4 --lazy-address /tmp/userfault.socket \
 --page-client --address source-system-ip --port 27

 and

 criu restore -D /tmp/4 -j --lazy-pages \
 --lazy-address /tmp/userfault.socket

Signed-off-by: Adrian Reber <areber@redhat.com>