cg: Add ability to dump custom cgroup properties and skip controllers
We have some common predefined properties such as "cpuset.cpus" and etc gathered
in @cgp_predefined set, but there might be situation when only predefined ones
are not enough, so add ability to specify additional properties via --cgroup-props
and/or --cgroup-props-file options.

For example one may pass

  --cgroup-props "\"cpu\": [\"cpu.shares\", \"cpu.cfs_period_us\"]"

to dump custom properties for cpu controller.

The description is implemented in almost valid yaml, probably we will
need to support the various forms, but oneline is enough for now.

Note, that once controllers and properties are defined, the controllers
met in procfs output not matched ones specified in command line are
skipped, moreover all predefine properties are skipped as well, so
that such dynamic descriptor must cover all controllers and properties
needed to be carried inside image.

I didn't tested all this, so new version probably coming soon.
Still take a look, maybe something obvious come to mind.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>