mount: add support for external block devices
The idea is to allow to dump and restore mounts of
a specified block device.

Options:
dump:      --external dev[MAJOR:MINOR]:VAL
restore:   --external VAL:DEVPATH

If we find a mount with a specified block device, we
set its type into FSTYPE__AUTO and write VAL into
the "source" field.

VAL is replaced on DEVPATH on restore.

https://jira.sw.ru/browse/PSBM-39381

v2: use --ext-mount-map instead of --external on dump
v3: clean up
v4: use --external on dump and on restore

Signed-off-by: Andrew Vagin <avagin@virtuozzo.com>