criu: add make lazy support to crit
This enables crit to remove all memory pages from a checkpoint
directory which can be lazily restored using userfaultfd. This
changes the pagemap.img and pages.img to no longer contain pages
which can be handled by userfaultfd (MAP_PRIVATE && MAP_ANON).

Usage:

 $ crit/crit make-lazy /tmp/4/ /tmp/5
 $ du -hs /tmp/4 /tmp/5
 201M	/tmp/4
 116K	/tmp/5

The checkpoint in /tmp/5 can be used by the actual restore process
and the checkpoint in /tmp/4 (with all memory pages) can be used
by the uffd daemon which then transfers the pages into the restored
on demand.

Signed-off-by: Adrian Reber <areber@redhat.com>