crit: merge criu-coredump
It was originally developed as a "crit coredump" action, but it
was decided to move it into a separate project about a year ago.
Several month ago there was a discussion about moving it back to
criu repo, so here is a patch to consider.

It gone through some RFC and fixing to make it work with criu 2.x.

Description from an original set:

This set itroduces a new CRIT feature called "core-dump". The name
is pretty much self-explanatory, it allows generating core dumps
from criu images. Core dumps are readable and usable by gdb.
The structure is something in the middle of kernel-generated
and gdb-generated core dumps.

This is a first iteration of core-dump patch series. It introduces
ability to generate gdb-operatable core dump only for x86_64 machine.
To support other architectures there will be a set to introduce proper
structures for notes section of the core dump. It is also worth noting,
that we still leave some notes(SIGINFO primarily) not fullfilled, as
I still need to figure out what to put in there in our case, where no
signal is involved in triggering core dump generation and other
caveats like zeroed vvar(we have problems with it withing CRIU itself
as well as gdb has some difficalties with it) and vsyscall vmas.
One can already use produced core dump with gdb, there is also
a quick simple demo https://asciinema.org/a/18936 .

Signed-off-by: Ruslan Kuprieiev <kupruser@gmail.com>