mount: create a clean mount only when it's required
"clean mount" is a copy of target mount without child mounts.

Currently a clean mount is created when a target mount has children.
In this case a target path MAY be overmounted.

In this patch, a clean mount is created only if a target path is
overmounted. For that we enumerate all children and check that they
are not mounted over a target path.

Signed-off-by: Andrew Vagin <avagin@virtuozzo.com>