vma: Don't set vma_area->vm_file_fd to -1 in alloc
The vma_area contains union-ed fields for dump and restore. The
generic alloc_vma_area() sets -1 to vm_file_fd which is a) dump
field and b) intersects with some restore fields. Thus at restore
time the structure that is expected to be all zeroed is not such.

Fix this and set vm_file_fd to -1 only when really needed.

Signed-off-by: Pavel Emelyanov <xemul@virtuozzo.com>
Acked-by: Cyrill Gorcunov <gorcunov@openvz.org>