unix: Resolve senders of packets in receive queue of DGRAM socket
Determine unique senders of packets in receive queue. If sender
is the only or if queue is empty, try to assign a queuer.
If there are several senders or if it's impossible to assign
a queuer, request the sender fle to be sent to task, who restores
the socket.

In resolve_unix_peers(), there are the second iteration thru
unix_sockets. It's need to set peers of rest sockets: sockets
who are not queuer of someone. I do not merge it in the single
iteration, because of optimization reasons. The most sockets
should have peers set after the first iteration, so to call
find_unix_sk_by_ino() one more time is excess.

v5: Use fle to send promious queue's senders.
Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>