mount: Forced mount unmounted binfmt_misc to do not lost its content
Umount does not remove binfmt_misc content. If it's mounted once again,
the same entries remain registered.

Criu does not dump content of umounted binfmt_misc. So, after C/R we
lose it at all.

This patch forces mounting of unmounted binfmt_misc before we collect
mountpoints. If it's unmounted, we mount it back and add this mount
to the list of forced mounted mountpoints. Next patch need this
list to mark the mount in special way in dump image.

Signed-off-by: Kirill Tkhai <ktkhai@virtuozzo.com>