Fix dumping of fds pointing to the /proc folder itself
CRIU fails to dump processes that have a file descriptor pointing
to the /proc folder, because check_path_remap returns error code
in that case. This is a bit odd since we allow dumping of
descriptors that are pointing to other folders/files in procfs.
To fix this, just return 0 if we know that we are looking at the
/proc folder itself.

Signed-off-by: Andrei Tuicu <andrei.tuicu@gmail.com>