gc: mock implementation of new gc command
gc is a new CRIU command. Its aim is to remove objects left
in the system after dump was performed.
Its main use case is to cleanup on node where dump was
performed but restore isn't planned.

CRIU gc accepts dump dir path as an input and removes:
1. Link remaps created for ps tree
2. Netfilter rules that drop ps tree tcp packets
Currently this is all that CRIU dump leaves in the system

--show option changes gc command behavior. It only
shows what was left in the system without objects removal.

Signed-off-by: Eugene Batalov <eabatalov89@gmail.com>