mount: open a root directory for mount namesapces via the root task
We use openat() syscall to open files in various mount namespaces.
Recently we found that if a path contains absolute symlinks, openat()
syscall resolves them relative to the current root, but they has to be
resolved relative to the root of the target namespace, so we need to
change root before restoring a file descriprot.

But for that we need to rework a method how we open a root directory
for a specified mount namespace. Currently we open /proc/pid/root for
one of processes from this namespace. If a process will change a root
directory, this way will not work.

In this patch we open all namespaces in the root task and then any
process will be able to open one of these descriptors via /proc/pid/fd.

Signed-off-by: Andrew Vagin <avagin@virtuozzo.com>