compel: Initial commit for rework
We're going to rework compel to support parasite
engine and plugin structure so that CRIU would simply
use the scaffolds.

The plugins gonna provide syscalls, in/out routunes,
file descriptor transfer and such, just like we currently
do in CRIU itself.

Here we shuffle the directories and fixup various things.

Note the @hostprogs are now left, probably will bring it back
one day. For same reason tests won't work for now.

Moreover since we're goig to ship compel/libcompel together
with criu and same time we are to *use* compel during CRIU
building procedure, the @hostprogs mode likely to be removed
completely and compel is to build two times in case of
cross-compile: once when building criu and second when
prepare its shipping. This is in TODO list currently.

Signed-off-by: Cyrill Gorcunov <gorcunov@openvz.org>